pyproject.toml
[project]
name = "eskro-api"
version = "0.1.0"
description = ""
authors = [
    {name = "terlyne",email = "sheyynovd@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi[all] (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "sqlalchemy[asyncio] (>=2.0.43,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "fastapi-mail (>=1.5.0,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "aiofiles (>=24.1.0,<25.0.0)",
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
alembic = "^1.16.5"
black = "^25.1.0"

